reorganize tests for rise
