+++ tests/methods/test_lime_image.py
@@ -3,7 +3,7 @@ from unittest import TestCase
 import numpy as np
 import dianna
 from dianna.methods.lime_image import LIMEImage
-from tests.test_onnx_runner import generate_data
+from tests.methods.test_onnx_runner import generate_data
 from tests.utils import run_model
